move markers and style to dedicated map endpoint
