+++ server/src/controllers/index.ts
@@ -13,6 +13,7 @@ import { FaceController } from 'src/controllers/face.controller';
 import { ReportController } from 'src/controllers/file-report.controller';
 import { JobController } from 'src/controllers/job.controller';
 import { LibraryController } from 'src/controllers/library.controller';
+import { MapController } from 'src/controllers/map.controller';
 import { MemoryController } from 'src/controllers/memory.controller';
 import { OAuthController } from 'src/controllers/oauth.controller';
 import { PartnerController } from 'src/controllers/partner.controller';
@@ -45,6 +46,7 @@ export const controllers = [
   FaceController,
   JobController,
   LibraryController,
+  MapController,
   MemoryController,
   OAuthController,
   PartnerController,
