+++ server/src/interfaces/map.interface.ts
@@ -1,0 +1,5 @@
+export const IMapRepository = 'IMapRepository';
+
+export interface IMapRepository {
+  fetchStyle(url: string): Promise<any>;
+}
