+++ server/src/services/index.ts
@@ -13,6 +13,7 @@ import { DownloadService } from 'src/services/download.service';
 import { DuplicateService } from 'src/services/duplicate.service';
 import { JobService } from 'src/services/job.service';
 import { LibraryService } from 'src/services/library.service';
+import { MapService } from 'src/services/map.service';
 import { MediaService } from 'src/services/media.service';
 import { MemoryService } from 'src/services/memory.service';
 import { MetadataService } from 'src/services/metadata.service';
@@ -38,11 +39,10 @@ import { UserService } from 'src/services/user.service';
 import { VersionService } from 'src/services/version.service';
 
 export const services = [
-  ApiService,
-  MicroservicesService,
   APIKeyService,
   ActivityService,
   AlbumService,
+  ApiService,
   AssetMediaService,
   AssetService,
   AssetServiceV1,
@@ -54,9 +54,11 @@ export const services = [
   DuplicateService,
   JobService,
   LibraryService,
+  MapService,
   MediaService,
   MemoryService,
   MetadataService,
+  MicroservicesService,
   NotificationService,
   PartnerService,
   PersonService,
@@ -73,7 +75,7 @@ export const services = [
   TagService,
   TimelineService,
   TrashService,
-  UserService,
   UserAdminService,
+  UserService,
   VersionService,
 ];
