+++ server/test/repositories/map.repository.mock.ts
@@ -1,0 +1,8 @@
+import { IMapRepository } from 'src/interfaces/map.interface';
+import { Mocked } from 'vitest';
+
+export const newMapRepositoryMock = (): Mocked<IMapRepository> => {
+  return {
+    fetchStyle: vitest.fn(),
+  };
+};
